Derby dependency: Removed the "tomcat" profile so that the derby dependency is no longer added in any war file. For the same raison, removed completly any reference to geotk-epsg-javadb, which should not be allowed at all in the Constellation project.
The Derby and PostgreSQL dependencies are now declared in the root pom.xml with the 'test' scope for every modules.
Only a few modules override this scope with an other one (by default 'runtime').

Also removed the explicit declaration of geoapi dependency, because this dependency
was already declared in the root pom.xml for every files.